# backend/API/openapi.yml
openapi: 3.1.0
info:
  title: Training Management API
  description: API for managing training sessions, students, trainers, schedules, and plans
  version: v1.0.0
servers:
- url: http://localhost:8080
  description: Local development server
security:
- bearerAuth: []
paths:
...
  /students/{studentId}:
    get:
      tags:
      - people
      operationId: findStudentById
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StudentResponseDTO'
    put:
      tags:
      - people
      operationId: updateStudent
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StudentResponseDTO'
    delete:
      tags:
      - people
      operationId: deleteStudent
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
  /students/{studentId}/restore:
    patch:
      tags:
      - people
      operationId: restoreStudent
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StudentResponseDTO'
  # ============================================
  # Training Resources Domain - Exercises
  # ============================================
...
  /students:
    get:
      tags:
      - people
      operationId: findAllStudents
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: gender
        in: query
        required: false
        schema:
          type: string
          pattern: (?i)^[MFO]$
      - name: profession
        in: query
        required: false
        schema:
          type: string
      - name: minAge
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 150
          minimum: 0
      - name: maxAge
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 150
          minimum: 0
      - name: registrationPeriodMinDate
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: registrationPeriodMaxDate
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: includeInactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageStudentResponseDTO'
    post:
      tags:
      - people
      operationId: createStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/StudentResponseDTO'
...
  /students/{studentId}/commitments/sessions/{sessionSeriesId}:
    post:
      tags:
      - scheduling
      operationId: updateCommitment
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sessionSeriesId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCommitmentRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommitmentDetailResponseDTO'
  /students/{studentId}/commitments/bulk:
    post:
      tags:
      - scheduling
      operationId: bulkUpdateCommitments
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCommitmentRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommitmentDetailResponseDTO'
...
  /students/{studentId}/commitments:
    get:
      tags:
      - scheduling
      operationId: getStudentCommitments
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommitmentDetailResponseDTO'
  /students/{studentId}/commitments/sessions/{sessionSeriesId}/status:
    get:
      tags:
      - scheduling
      operationId: getCurrentCommitmentStatus
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sessionSeriesId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: timestamp
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                enum:
                - ATTENDING
                - NOT_ATTENDING
                - TENTATIVE
  /students/{studentId}/commitments/sessions/{sessionSeriesId}/history:
    get:
      tags:
      - scheduling
      operationId: getCommitmentHistory
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sessionSeriesId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommitmentDetailResponseDTO'
  /students/{studentId}/commitments/active:
    get:
      tags:
      - scheduling
      operationId: getCurrentActiveCommitments
      parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: timestamp
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommitmentDetailResponseDTO'
  /schedules/{sessionSeriesId}/commitments:
    get:
      tags:
      - scheduling
      operationId: getSessionSeriesCommitments
      parameters:
      - name: sessionSeriesId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommitmentDetailResponseDTO'
...
components:
  schemas:
...
    AnamnesisRequestDTO:
      type: object
      properties:
        medication:
          type: string
        isDoctorAwareOfPhysicalActivity:
          type: boolean
        favoritePhysicalActivity:
          type: string
        hasInsomnia:
          type: string
          enum:
          - 'yes'
          - 'no'
          - sometimes
        dietOrientedBy:
          type: string
        cardiacProblems:
          type: string
        hasHypertension:
          type: boolean
        chronicDiseases:
          type: string
        difficultiesInPhysicalActivities:
          type: string
        medicalOrientationsToAvoidPhysicalActivity:
          type: string
        surgeriesInTheLast12Months:
          type: string
        respiratoryProblems:
          type: string
        jointMuscularBackPain:
          type: string
        spinalDiscProblems:
          type: string
        diabetes:
          type: string
        smokingDuration:
          type: string
        alteredCholesterol:
          type: boolean
        osteoporosisLocation:
          type: string
    PhysicalImpairmentRequestDTO:
      type: object
      properties:
        type:
          type: string
          enum:
          - visual
          - auditory
          - motor
          - intellectual
          - other
        name:
          type: string
          maxLength: 255
          minLength: 0
        observations:
          type: string
      required:
      - name
      - type
    StudentRequestDTO:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        name:
          type: string
          maxLength: 100
          minLength: 0
        surname:
          type: string
          maxLength: 100
          minLength: 0
        gender:
          type: string
          enum:
          - M
          - F
          - O
        birthDate:
          type: string
          format: date
        phone:
          type: string
          maxLength: 20
          minLength: 0
        profession:
          type: string
          maxLength: 100
          minLength: 0
        street:
          type: string
          maxLength: 255
          minLength: 0
        number:
          type: string
          maxLength: 10
          minLength: 0
        complement:
          type: string
          maxLength: 100
          minLength: 0
        neighborhood:
          type: string
          maxLength: 100
          minLength: 0
        cep:
          type: string
          maxLength: 10
          minLength: 0
        emergencyContactName:
          type: string
          maxLength: 100
          minLength: 0
        emergencyContactPhone:
          type: string
          maxLength: 20
          minLength: 0
        emergencyContactRelationship:
          type: string
          maxLength: 50
          minLength: 0
        objectives:
          type: string
        observations:
          type: string
        anamnesis:
          $ref: '#/components/schemas/AnamnesisRequestDTO'
        physicalImpairments:
          type: array
          items:
            $ref: '#/components/schemas/PhysicalImpairmentRequestDTO'
      required:
      - birthDate
      - email
      - gender
      - name
      - surname
    AnamnesisResponseDTO:
      type: object
      properties:
        medication:
          type: string
        isDoctorAwareOfPhysicalActivity:
          type: boolean
        favoritePhysicalActivity:
          type: string
        hasInsomnia:
          type: string
          enum:
          - 'yes'
          - 'no'
          - sometimes
        dietOrientedBy:
          type: string
        cardiacProblems:
          type: string
        hasHypertension:
          type: boolean
        chronicDiseases:
          type: string
        difficultiesInPhysicalActivities:
          type: string
        medicalOrientationsToAvoidPhysicalActivity:
          type: string
        surgeriesInTheLast12Months:
          type: string
        respiratoryProblems:
          type: string
        jointMuscularBackPain:
          type: string
        spinalDiscProblems:
          type: string
        diabetes:
          type: string
        smokingDuration:
          type: string
        alteredCholesterol:
          type: boolean
        osteoporosisLocation:
          type: string
    PhysicalImpairmentResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - visual
          - auditory
          - motor
          - intellectual
          - other
        name:
          type: string
        observations:
          type: string
    StudentResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        surname:
          type: string
        gender:
          type: string
          enum:
          - M
          - F
          - O
        birthDate:
          type: string
          format: date
        phone:
          type: string
        profession:
          type: string
        street:
          type: string
        number:
          type: string
        complement:
          type: string
        neighborhood:
          type: string
        cep:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        emergencyContactRelationship:
          type: string
        objectives:
          type: string
        observations:
          type: string
        registrationDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
        anamnesis:
          $ref: '#/components/schemas/AnamnesisResponseDTO'
        physicalImpairments:
          type: array
          items:
            $ref: '#/components/schemas/PhysicalImpairmentResponseDTO'
...
    StudentCommitmentRequestDTO:
      type: object
      properties:
        commitmentStatus:
          type: string
          enum:
          - ATTENDING
          - NOT_ATTENDING
          - TENTATIVE
        effectiveFromTimestamp:
          type: string
          format: date-time
      required:
      - commitmentStatus
    CommitmentDetailResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        studentName:
          type: string
        sessionSeriesId:
          type: string
          format: uuid
        seriesName:
          type: string
        commitmentStatus:
          type: string
          enum:
          - ATTENDING
          - NOT_ATTENDING
          - TENTATIVE
        effectiveFromTimestamp:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    BulkCommitmentRequestDTO:
      type: object
      properties:
        sessionSeriesIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
        commitmentStatus:
          type: string
          enum:
          - ATTENDING
          - NOT_ATTENDING
          - TENTATIVE
        effectiveFromTimestamp:
          type: string
          format: date-time
      required:
      - commitmentStatus
      - sessionSeriesIds
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageStudentResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/StudentResponseDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT